pyproject.toml
[project]
name = "mensa-mcp"
version = "0.1.0"
description = "MCP server for university restaurant menus"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "httpx>=0.28.1",
    "mcp[cli]>=1.27.0",
    "pyyaml>=6.0.3",
    "sse-starlette>=3.3.4",
    "starlette>=1.0.0",
    "uvicorn>=0.44.0",
]

[project.scripts]
mensa-mcp = "mensa_mcp:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mensa_mcp"]
